add anthropic generator tests
